Move cert issuance code to pki module
