+++ pkg/pki/BUILD.bazel
@@ -3,8 +3,10 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = [
+        "cert_utils.go",
         "certificate.go",
         "csr.go",
+        "issue.go",
         "privatekey.go",
         "sshkey.go",
     ],
