+++ upup/pkg/fi/fitasks/BUILD.bazel
@@ -3,7 +3,6 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = [
-        "cert_utils.go",
         "keypair.go",
         "keypair_fitask.go",
         "managedfile.go",
